Clean pkg configurations
