+++ openstack/autoscaling/v1/configurations/delete.go
@@ -6,3 +6,7 @@ func Delete(client *golangsdk.ServiceClient, id string) (r DeleteResult) {
 	_, r.Err = client.Delete(client.ServiceURL("scaling_configuration", id), nil)
 	return
 }
+
+type DeleteResult struct {
+	golangsdk.ErrResult
+}
